requirements.txt
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
idna==2.10
MarkupSafe==1.1.1
pbkdf2==1.3
prettytable==3.5.0
pyaes==1.6.1
PyJWT==2.3.0
pytz==2021.3
requests==2.31.0
twilio==7.3.2
urllib3==2.2.1
wcwidth==0.2.5
pyasn1==0.6.0
cryptography==42.0.5
python-telegram-bot==21.0.1